payment get id

+++ src/main/java/com/math/mathcha/dto/request/RechargeRequestDTO.java
@@ -1,11 +1,13 @@
 package com.math.mathcha.dto.request;
 
 
+import com.math.mathcha.entity.Course;
+import com.math.mathcha.entity.Student;
 import lombok.Data;
 
 @Data
 public class RechargeRequestDTO {
     String amount;
-    int studentId;
-    int courseId;
+    Long studentId;
+    Long courseId;
 }
